chore: restructure agent into nodes and refactor extract
Move and reorganise the agent code into node-style packages and simplify alert extraction.

- Remove legacy app/agent/investigation.py and add app/agent/nodes/investigate/agent.py implementing the ReAct investigation loop (preserves behaviour but relies on helpers from nodes.investigate.loop and nodes.investigate.tools).
- Move app/agent/extract.py to app/agent/nodes/extract_alert/node.py and refactor extraction logic: introduce AlertDetails pydantic model, clearer prompt text, separate helper functions (_log_raw_alert, _handle_noise, _render_alert_summary, _build_alert_updates, _dict_value, _first_value, _needs_full_json_prompt, _slack_reaction_context), improved raw_alert enrichment and incident_window handling, and safer Slack reaction context extraction.
- Add package __init__.py files for app.agent.nodes.extract_alert and app.agent.nodes.investigate and update app/agent/__init__.py to import from the new investigate node path.
- Wire up existing tool/loop helpers (duplicate suppression, seed calls, tool context building, etc.) by moving/using helpers from app.agent.nodes.investigate.tools and loop modules.
- Minor logging and tracker updates; update fields reported on tracker completion.

Tests and other modules were adjusted to accommodate the new module paths and refactored helper APIs.
