+++ app/agent/__init__.py
@@ -1,6 +1,6 @@
 """Connected ReAct agent loop for raw-alert-first investigations."""
 
 from app.agent.chat import ChatAgent
-from app.agent.investigation import ConnectedInvestigationAgent, InvestigationAgent
+from app.agent.nodes.investigate import ConnectedInvestigationAgent, InvestigationAgent
 
 __all__ = ["ConnectedInvestigationAgent", "InvestigationAgent", "ChatAgent"]
