+++ app/agent/nodes/extract_alert/__init__.py
@@ -1,0 +1,5 @@
+"""Extract alert node — classify and parse raw alerts into structured state."""
+
+from app.agent.nodes.extract_alert.node import extract_alert
+
+__all__ = ["extract_alert"]
