+++ app/agent/nodes/investigate/__init__.py
@@ -1,0 +1,5 @@
+"""Investigate node — connected ReAct investigation agent."""
+
+from app.agent.nodes.investigate.agent import ConnectedInvestigationAgent, InvestigationAgent
+
+__all__ = ["ConnectedInvestigationAgent", "InvestigationAgent"]
