+++ app/agent/nodes/investigate/tools.py
@@ -1,0 +1,225 @@
+"""Tool resolution, seeding, and evidence helpers for the investigate node."""
+
+from __future__ import annotations
+
+from typing import Any
+
+from app.agent.tool_loop import _public_tool_input
+from app.services.agent_llm_client import ToolCall
+from app.tools.registered_tool import RegisteredTool
+from app.tools.registry import get_registered_tools
+from app.utils.tool_trace import redact_sensitive
+
+# Maps alert_source → tool source keys. Tools from these sources are auto-called
+# before the LLM loop starts when the alert source is known.
+_ALERT_SOURCE_TO_TOOL_SOURCES: dict[str, list[str]] = {
+    "grafana": ["grafana"],
+    "datadog": ["datadog"],
+    "cloudwatch": ["cloudwatch"],
+    "eks": ["eks"],
+    "alertmanager": ["grafana", "cloudwatch"],
+    "sentry": ["sentry"],
+    "honeycomb": ["honeycomb"],
+    "coralogix": ["coralogix"],
+    "airflow": ["airflow"],
+    "hermes": ["hermes"],
+    "kafka": ["kafka"],
+    "postgresql": ["postgresql"],
+    "mysql": ["mysql"],
+    "mariadb": ["mariadb"],
+    "mongodb": ["mongodb", "mongodb_atlas"],
+    "redis": ["redis"],
+    "snowflake": ["snowflake"],
+    "clickhouse": ["clickhouse"],
+    "dagster": ["dagster"],
+    "rabbitmq": ["rabbitmq"],
+    "supabase": ["supabase"],
+    "opensearch": ["opensearch"],
+    "openobserve": ["openobserve"],
+    "betterstack": ["betterstack"],
+    "azure": ["azure", "azure_sql"],
+    "splunk": ["splunk"],
+    "signoz": ["signoz"],
+    "jenkins": ["jenkins"],
+    "tempo": ["tempo"],
+}
+
+# Consecutive iterations made up ENTIRELY of duplicate (already-seen) tool calls
+# that we tolerate before forcing the agent to conclude.
+MAX_STAGNANT_ITERATIONS = 2
+
+# Injected as a user turn once the agent starts repeating itself.
+STAGNATION_NUDGE = (
+    "You are repeating tool calls you already made, so they return no new "
+    "information and the investigation is not progressing. Stop calling tools and "
+    "write your final diagnosis from the evidence already gathered: root cause, "
+    "root cause category, supporting evidence, validated and non-validated claims, "
+    "remediation steps, and a validity score. If the evidence is insufficient to "
+    "determine a root cause, say so explicitly and use a low validity score."
+)
+
+
+def get_available_tools(resolved_integrations: dict[str, Any]) -> list[RegisteredTool]:
+    available_sources = availability_view(resolved_integrations)
+    return [t for t in get_registered_tools("investigation") if t.is_available(available_sources)]
+
+
+def availability_view(resolved_integrations: dict[str, Any]) -> dict[str, Any]:
+    """Adapt resolved integration configs to the legacy tool availability contract."""
+    view: dict[str, Any] = {}
+    for key, value in resolved_integrations.items():
+        if key.startswith("_") or not isinstance(value, dict) or not value:
+            view[key] = value
+            continue
+        item = dict(value)
+        item.setdefault("connection_verified", True)
+        view[key] = item
+    return view
+
+
+def build_connected_tool_context(
+    resolved_integrations: dict[str, Any],
+    tools: list[RegisteredTool],
+) -> dict[str, Any]:
+    from app.integrations.registry import family_key
+
+    connected_integrations = sorted(
+        key
+        for key, value in resolved_integrations.items()
+        if not key.startswith("_") and isinstance(value, dict) and value
+    )
+    connected_families = {family_key(key) for key in connected_integrations}
+
+    sources: dict[str, dict[str, Any]] = {}
+    for tool in sorted(tools, key=lambda item: (str(item.source), item.name)):
+        source = str(tool.source)
+        source_info = sources.setdefault(
+            source,
+            {
+                "connected": source in connected_integrations
+                or family_key(source) in connected_families,
+                "tools": [],
+            },
+        )
+        source_info["tools"].append(tool.name)
+
+    return {
+        "connected_integrations": connected_integrations,
+        "available_sources": sources,
+        "available_action_names": [tool.name for tool in sorted(tools, key=lambda item: item.name)],
+    }
+
+
+def build_seed_calls(
+    state: dict[str, Any],
+    tools: list[RegisteredTool],
+    llm: Any,
+) -> list[ToolCall]:
+    """Return tool calls to run before the LLM loop based on the alert source."""
+    alert_source = get_alert_source(state)
+    if not alert_source:
+        return []
+
+    target_sources = set(_ALERT_SOURCE_TO_TOOL_SOURCES.get(alert_source, []))
+    if not target_sources:
+        return []
+
+    resolved = state.get("resolved_integrations") or {}
+    seed_tools = [t for t in tools if str(t.source) in target_sources]
+    if not seed_tools:
+        return []
+
+    from app.services.agent_llm_client import BedrockConverseAgentClient
+    from app.services.bedrock_converse import new_tool_use_id
+
+    use_converse_ids = isinstance(llm, BedrockConverseAgentClient)
+    calls: list[ToolCall] = []
+    for tool in seed_tools:
+        try:
+            injected = tool.extract_params(resolved)
+        except Exception:
+            injected = {}
+        tool_id = new_tool_use_id() if use_converse_ids else f"seed_{tool.name}"
+        calls.append(ToolCall(id=tool_id, name=tool.name, input=_public_tool_input(injected)))
+
+    return calls
+
+
+def get_alert_source(state: dict[str, Any]) -> str:
+    source = str(state.get("alert_source") or "").lower().strip()
+    if source:
+        return source
+    raw = state.get("raw_alert")
+    if isinstance(raw, dict):
+        source = str(raw.get("alert_source") or "").lower().strip()
+        if source:
+            return source
+        labels = raw.get("commonLabels") or raw.get("labels") or {}
+        if isinstance(labels, dict) and (
+            labels.get("grafana_folder") or labels.get("datasource_uid")
+        ):
+            return "grafana"
+        ext_url = raw.get("externalURL", "")
+        if isinstance(ext_url, str) and "grafana" in ext_url.lower():
+            return "grafana"
+    return ""
+
+
+def tool_event_payload(tc: ToolCall, *, output: Any | None = None) -> dict[str, Any]:
+    payload: dict[str, Any] = {
+        "id": tc.id,
+        "name": tc.name,
+        "input": redact_sensitive(tc.input),
+    }
+    if output is not None:
+        payload["output"] = redact_sensitive(output)
+    return payload
+
+
+def merge_tool_evidence(
+    evidence: dict[str, Any],
+    tool_name: str,
+    output: Any,
+    tool_input: dict[str, Any],
+) -> None:
+    """Store raw tool output and the legacy report-facing evidence keys."""
+    evidence[tool_name] = output
+    tool_outputs = evidence.setdefault("tool_outputs", [])
+    if isinstance(tool_outputs, list):
+        tool_outputs.append(
+            {
+                "tool_name": tool_name,
+                "tool_args": redact_sensitive(tool_input),
+                "data": redact_sensitive(output),
+            }
+        )
+
+    if not isinstance(output, dict):
+        return
+
+    if tool_name == "query_grafana_logs":
+        evidence["grafana_logs"] = output.get("logs", [])
+        evidence["grafana_error_logs"] = output.get("error_logs", [])
+        evidence["grafana_logs_query"] = output.get("query", "")
+        evidence["grafana_logs_service"] = output.get("service_name", "")
+        return
+
+    if tool_name == "query_grafana_metrics":
+        metric_name = str(output.get("metric_name") or tool_input.get("metric_name") or "")
+        metric_results = evidence.setdefault("grafana_metric_results", {})
+        if isinstance(metric_results, dict) and metric_name:
+            metric_results[metric_name] = output
+        evidence["grafana_metrics"] = output.get("metrics", [])
+        return
+
+    if tool_name == "query_grafana_traces":
+        evidence["grafana_traces"] = output.get("traces", [])
+        evidence["grafana_pipeline_spans"] = output.get("pipeline_spans", [])
+        return
+
+    if tool_name == "query_grafana_alert_rules":
+        evidence["grafana_alert_rules"] = output.get("rules", [])
+        return
+
+    if tool_name == "query_grafana_service_names":
+        evidence["grafana_service_names"] = output.get("service_names", [])
