+++ tests/benchmarks/cloudopsbench/bench_agent.py
@@ -1,6 +1,6 @@
 """CloudOpsBench-specific investigation agent.
 
-Subclasses :class:`app.agent.investigation.ConnectedInvestigationAgent` to
+Subclasses :class:`app.agent.nodes.investigate.ConnectedInvestigationAgent` to
 enforce a minimum-tool-call floor before the agent is allowed to conclude.
 Production code is untouched — bench-only termination behavior lives here.
 
@@ -30,7 +30,7 @@ import logging
 import os
 from typing import Any, ClassVar
 
-from app.agent.investigation import ConnectedInvestigationAgent
+from app.agent.nodes.investigate import ConnectedInvestigationAgent
 from app.tools.registered_tool import RegisteredTool
 
 logger = logging.getLogger(__name__)
