+++ tests/synthetic/test_signoz_scenario.py
@@ -8,7 +8,7 @@ from __future__ import annotations
 
 from typing import Any
 
-from app.agent.investigation import _ALERT_SOURCE_TO_TOOL_SOURCES
+from app.agent.nodes.investigate.tools import _ALERT_SOURCE_TO_TOOL_SOURCES
 from app.tools.SignozLogsTool import query_signoz_logs
 from app.tools.SignozMetricsTool import query_signoz_metrics
 from app.tools.SignozTracesTool import query_signoz_traces
