+++ tests/synthetic/test_tempo_scenario.py
@@ -12,7 +12,7 @@ import pytest
 
 pytestmark = pytest.mark.synthetic
 
-from app.agent.investigation import _ALERT_SOURCE_TO_TOOL_SOURCES
+from app.agent.nodes.investigate.tools import _ALERT_SOURCE_TO_TOOL_SOURCES
 from app.tools.TempoTool import query_tempo
